src/backend/requirements.txt
# command line parsing
click

# formatting
black

# pywebview
pywebview
pyobjc

# networking
requests
